Changes all coats and jackets with buttons to be subtypes of /obj/item/clothing/suit/storage/toggle and changes the way the toggle verb functions to be dependent on defined icon states within the item
This means that adding a new type of jacket that toggles will be much easier, as instead of having to duplicate the entire old function and hardcode icon states for open and closed, you can just add in a new item as a subtype of suit/storage/toggle and define the icon states there. Also makes on-the-fly modifying of icons easier.